Remove collection changes for followup PR
