--- app/components/collections/edit/form_component.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'collections/form_fields', locals: { collection_form: @collection_form } %>
