+++ app/components/dashboard/show/collections_list_component.html.erb
@@ -1,10 +1,7 @@
 <%= render Elements::HeaderComponent.new(level: :h3, text: label, classes: 'mb-4') %>
-<%= render Elements::ButtonLinkComponent.new(link: new_collection_path, label: 'Create a new collection', variant: :primary) %>
 
 <% collections.each do |collection| %>
-  <%= link_to "/collections/#{collection.druid}" do %>
-    <%= render Elements::HeaderComponent.new(level: :h4, text: collection.title) %>
-  <% end %>
+  <%= render Elements::HeaderComponent.new(level: :h4, text: collection.title) %>
   <%= render Dashboard::Show::WorksListComponent.new(collection:) %>
   <%= render Elements::ButtonLinkComponent.new(link: new_work_path(collection_druid: collection.druid), label: 'Deposit to this collection', variant: :primary) %>
 <% end %>
