+++ app/components/elements/tab_list_component.html.erb
@@ -1,5 +1,5 @@
 <%= tag.div class: classes, 'data-controller': 'tab-error' do %>
-  <%= tag.div class: tab_classes, role: 'tablist', 'aria-orientation': orientation do |component| %>
+  <%= tag.div class: tab_classes, role: 'tablist', 'aria-orientation': 'vertical' do |component| %>
     <% tabs.each do |tab| %>
       <%= tab %>
     <% end %>
