--- app/components/works/edit/form_component.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'works/form_fields', locals: { work_form: @work_form, content: @files } %>
