--- app/views/collections/show.html.erb
@@ -1,15 +1,0 @@
-<%= render Elements::HeaderComponent.new(level: :h1, text: @collection_form.title, classes: 'mb-4 d-inline-block') %>
-<span class="ms-4 fst-italic status"><%= @status_presenter.status_message %></span>
-<%= render Elements::ButtonLinkComponent.new(link: edit_collection_path(@collection_form.druid), label: 'Edit or deposit', classes: 'float-end', variant: :'outline-primary') if @status_presenter.editable? %>
-
-<%= render Elements::Tables::TableComponent.new(id: 'title-table', classes: 'table-work mb-5', label: 'Title') do |component| %>
-  <% component.with_row(label: 'Title', values: [@collection_form.title]) %>
-<% end %>
-
-<%= render Elements::Tables::TableComponent.new(id: 'description-table', classes: 'table-work mb-5', label: 'Description') do |component| %>
-  <% component.with_row(label: 'Description', values: [@collection_form.description]) %>
-<% end %>
-
-<%= render Elements::Tables::TableComponent.new(id: 'related-content-table', classes: 'table-work mb-5', label: 'Related content') do |component| %>
-  <% component.with_row(label: 'Related links', values: @collection_form.related_links_attributes.map { |related_link| link_to(related_link.text || related_link.url, related_link.url) }) %>
-<% end %>
