--- app/views/collections/wait.html.erb
@@ -1,2 +1,0 @@
-<%= turbo_stream_from 'wait', params[:id] %>
-<%= render Elements::SpinnerComponent.new(message: 'Processing...', variant: :info, classes: 'my-2') %>
