+++ config/initializers/custom_serializers.rb
@@ -1,4 +1,3 @@
 # frozen_string_literal: true
 
 Rails.application.config.active_job.custom_serializers << WorkFormSerializer
-Rails.application.config.active_job.custom_serializers << CollectionFormSerializer
