+++ config/locales/en.yml
@@ -88,48 +88,3 @@ en:
         title:
           tab_label: 'Title'
           label: 'Title of deposit'
-  collections:
-    edit:
-      fields:
-        description:
-          label: 'Description'
-          help_text: >
-            Enter a summary statement (max 600 words) describing the collection that you are creating.
-            A detailed statement may improve the discovery of your collection in web searches.
-        title:
-          label: 'Title of collection'
-          help_text: >
-            A unique, descriptive title may improve discovery of your collection in web searches.
-        license:
-          label: 'License'
-          help_text: Assigning a default license may improve discovery of your collection in web searches.
-          license_terms_of_use: In addition to the license, the following Terms of Use will also be displayed on your PURL page.
-          terms_of_use: >
-            User agrees that, where applicable, content will not be used to identify or to otherwise infringe the privacy or
-            confidentiality rights of individuals. Content distributed via the Stanford Digital Repository may be subject to
-            additional license and use restrictions applied by the depositor.
-      errors:
-        cannot_be_edited: 'This collection cannot be edited.'
-        validation: >
-          Required fields have not been filled out. Click on the section(s) highlighted in red on
-          the left to correct the error(s) and then try depositing again.
-      panes:
-        description:
-          tab_label: 'Description'
-          label: 'Describe your collection'
-        deposit:
-          tab_label: 'Deposit'
-          label: 'Submit your collection'
-          help_text: >
-              Be sure you have completed all the required sections.
-        license:
-          tab_label: 'License'
-          label: 'License'
-          help_text: >
-            More text here.
-        related_content:
-          tab_label: 'Related content (optional)'
-          label: 'Related content'
-        title:
-          tab_label: 'Title'
-          label: 'Title of collection'
